Refactored code so that runPhase only takes a String code to specify which phase.
runPhase is also responsible for logging success and failure of processing each individual block. It also sets up the ThreadContext appropraitely.

None of the sub-components take a Context object and thus this was removed from setupComonent.

No sub-component requires access to entities outside standard CM so this is just passed to them (only the frame at this stage).

Simplified setup of subcomponents.


git-svn-id: https://svn.apache.org/repos/asf/avalon/trunk@12010 13f79535-47bb-0310-9956-ffa450edef68