Update to use LogEnabled rather than Loggable where possible.
This means that internally the Phoenix container uses LogEnabled and only uses Loggable when a hosted block still uses the old format.

Eventually the old interface will be deprecated and removed.

Unfortunately this required a backwards incompatible change in BlockContext so that we return a org.apache.avalon.framework.logger.Logger rather than a org.apache.log.Logger. This should not effect a large proportion of users.


git-svn-id: https://svn.apache.org/repos/asf/avalon/trunk@12357 13f79535-47bb-0310-9956-ffa450edef68