Attempted to fix some synchronization bugs.
It was possible (and likely) that stop()/dispose() would be ignored and the block would continue on running scheduled events.

It was also possible in very rare circumstances for a newly added event to be ignored if it had close to zero time before it was next scheduled to run. I believe it has been fixed.


git-svn-id: https://svn.apache.org/repos/asf/avalon/trunk@12454 13f79535-47bb-0310-9956-ffa450edef68