When requesting a Resources object for a specific class/package it is likely that the resources will be loaded from same ClassLoader as specified class/package.
As it stands now this will fail if the resources are loaded when ContextClassLoader is not set to a class that can access passed in class. This is common when you have a component call back into container - if container has not loaded resources and tries to (in context of components callback) then this will fail.


git-svn-id: https://svn.apache.org/repos/asf/avalon/trunk@12653 13f79535-47bb-0310-9956-ffa450edef68