Fix up the build script situation. Each component and example should now be buildable via the scripts on Unix and Windows.
There is now a main build.sh and build.bat in the root directory, which does
all the hard work of finding and invoking Ant. Then, for all other
sub-projects, there are small, simple build scripts which just invoke the root
build script.

The advantage of this is that all the high-maintenance shell and batch script
code is kept and maintained in one place. With the breakup of the source tree
into components, and subsequent proliferation of build scripts, this has become
more important.

Also added a bit to BUILDING.txt for people building from CVS.


git-svn-id: https://svn.apache.org/repos/asf/avalon/trunk@13025 13f79535-47bb-0310-9956-ffa450edef68