Initial commit of a small component 'viewer' tool I wrote to test out the different initialization policies a component can specify about it's respective ComponentHandler.
It provides a small gui with a combobox list of all registered components
this container has, and the ability fire a 'lookup' event on any one of
them.

(by tracing the log, it's possible to see the difference on a lookup() between
request/startup based inits).

Perhaps this might be useful for others, or maybe extended to become a
gui component testing tool, etc ? Ideas more than welcome.


git-svn-id: https://svn.apache.org/repos/asf/avalon/trunk@14295 13f79535-47bb-0310-9956-ffa450edef68