Here's a patch to allow for the setting of the Host attribute on the MX4J HttpAdaptor.  This adds an optional 'manager-adapter-host' attribute that can be configured in the MX4JSystemManager section of kernel.xml.
The reason for this patch is that the HttpAdaptor only binds to localhost by
default, thus leaving it inaccessible to remote machines.  The default
behaviour of this patch is to still bind to localhost, as before, but one
can now optionally specify a host alias for a public interface.

Submitted by: "Marc Tremblay" <marc.tremblay@wmode.com>


git-svn-id: https://svn.apache.org/repos/asf/avalon/trunk@14889 13f79535-47bb-0310-9956-ffa450edef68