Start providing basic support for dependencies that are either; * an array of services of a particular type * a map of services of a particular type
ie It is now possible for a block to declare that it will accept a map of "Foo" services and the assembler will be able to map 0-N different Foo services into map. An assembler could also map 0-N services of type Foo onto a dependency if it is an array dependency.

Array dependencies have postfix "[]" while maps have a postfix "#"


git-svn-id: https://svn.apache.org/repos/asf/avalon/trunk@15027 13f79535-47bb-0310-9956-ffa450edef68