First step of migration of documentation to Forrest. New docs dir:   jakarta-avalon-phoenix/src/documentation
New content dir:
  jakarta-avalon-phoenix/src/documentation/content

New resources and images dir:
  jakarta-avalon-phoenix/src/documentation/resources

To build the documentation download Forrest from CVS
   http://xml.apache.org/forrest/your-project.html#N10022
 install it as described here
   http://xml.apache.org/forrest/your-project.html#N10036

cd in the  jakarta-avalon-phoenix dir and run "forrest"

To see the site real-time from a local embedded webserver run "forrest run"
and point the browser to http://localhost:8888/
The sources to change are in ./build/webapp, you will need to synch them with
the actual docs manually.

Or run the Phoenix build.xml with target "site".


git-svn-id: https://svn.apache.org/repos/asf/avalon/trunk@15698 13f79535-47bb-0310-9956-ffa450edef68