From PeterD...
It does a bunch of things. Basically it adds support for "Partitions" which
are groupings of components that are operated on in concert. ie The listeners
in phoenix would be one partition and the blocks would be another. Nothing
particularly earth shattering and it essentially follows the design of HP's
CSF.

The patch also copies back the array/map dependency stuff from phoenix into
containerkit.

It also adds the notion of attributes to the metadata/assembly declarations.

It also vastly imprves the memory management not requiring so many
instantiations and dealing with nulls in some cases and defining empty arrays
in others.

It also vastly improves the support for handling BlockInfo style metadata,
including offering writing support.

Also implements proper mapping of phoenix style short names for schema (ie
relax-ng) ro uri based names (ie http://www.relaxng ....). Which makes it
easier to support a larger variety of schema languages to be used in the
future.


git-svn-id: https://svn.apache.org/repos/asf/avalon/trunk@15746 13f79535-47bb-0310-9956-ffa450edef68