Changes to remove the self-resetting behavior. The strategy did a reset on itself when it returned true from the isRotationNeeded() method - it didn't wait for a reset().
Also added some Javadoc explaining the known problems with the class.

A possible solution to the problem is to set m_currentSize to file.length() if
m_currentSize == 0 in the isRotationNeeded method, or by reset()-ing to the
size of the last message, but at the time I don't want to do such changes to
the code as we have a release coming up.


git-svn-id: https://svn.apache.org/repos/asf/avalon/trunk@16142 13f79535-47bb-0310-9956-ffa450edef68