Applied patch from Bruno Dumon:
Attached is a small patch for the XPathProcessorImpl from excalibur xmlutil.

It solves the problem that XPath expressions like "/a/b" don't return
any result, even if there are elements in the document conforming to
that XPath expression. (these kind of expressions were actually
throwing exceptions but those are silently catched -- something I'll
address at a later time).

The purpose of Xalan's PrefixResolver.handlesNullPrefixes() method
is not entirely clear to me, but returning false helps (which is also what
Xalan itself does in all of its PrefixResolver implementations).


git-svn-id: https://svn.apache.org/repos/asf/avalon/trunk@17098 13f79535-47bb-0310-9956-ffa450edef68