I've been doing some digging into this problem. The inital reason why a file url was included in the maven.xdocs.jsl was due to a bug back in Maven 8.  That bug was resolved in maven 9, however the workaround seems to still be present in the meta package build procedures.  I've tried to resolve this using the approach that Alexis suggested (and a dozen other variations) but this breaks under Windows.
Conclusion - there is a problem in variable resolution that is probably linked to ant properties verus jelly variable assignments in the context of reactor builds.

This update involves the shifting of all site related defintions from site/project.xml to the root directory under platform.xml (the same approach as used in the merlin package).  This ensures that we are always running from the same ${basedir) and that the properties file and meven.xml are the same when running a build and when running site generation.  Under this approach the defintion of the maven.xdoc.jsl  is simply "site/etc/site.jsl".


git-svn-id: https://svn.apache.org/repos/asf/avalon/trunk@18171 13f79535-47bb-0310-9956-ffa450edef68